Initial API commit, mostly copied from DuckASM
DuckASM was my internal API for code-generation in my JIT (DuckLogic).
The good thing is is that this project is open source!
The bad news is that the rest of DuckLogic isn't (yet).

This was the basic API I was using for static reflection,
with a little more documentation added on

I'll copy the implementation over soon